// src/common/enum/otpKey.enum.ts
//~ Assignment 19 ~//

export enum OTPKeyEnum {
  signUp = "signUp",
  forgotPassword = "forgotPassword",
}

export enum subjectEnum {
  signUp = "Confirm your email",
  resetPassword = "Reset your password",
}

export enum otpTemplateEnum {
  signUp = "complete your verification process",
  resetPassword = "reset your password",
}